HTML: Moved everything into their new tabs and added the Domains tab content.
CSS: Styled the new tabs and changed the styling of a lot of the content to better fit within the new Tabs style.

JS: Added the script for switching tab functionality.

Notes: Everything worked in the single html test file that had the css and script in-line. Hopefully it all works together in seperate files!